Include custom module Sortable in Tree class
Sortable contains the logic for a merge sort method, to be used when
building a Tree. A Tree object requires that the provided array is
sorted, but does not assume it to already be so.

When initialized, a Tree will set its root to be the return value of its
build_tree method, although it doesn't do that yet. Before starting on
the logic for build_tree, Sortable has been included and tested, as the
array provided to build_tree must be sorted. And so here we are!