Implement #build_tree for initialization
Tree objects initialize by calling #build_tree on a provided array. The
method itself assumes the array is sorted already on its own, but in the
initialize method #merge_sort is called, just to be sure.

Pretty_print is provided as a convenient and aesthetically-pleasing way
of viewing the tree in a terminal.