Implement #insert for adding new leaf nodes
This will look down the tree by checking if the current node is greater
or lesser in value to the one passed to the method. If the passed value
is less, move left, otherwise right, until you reach a nil value.

This can result in the tree becoming unbalanced, especially if #insert
is called multiple times. In the future there will be methods to check
if the tree is balanced or not, and to rebalance it with current values.