Implement #delete functionality
Deletes a specified value in the tree in one of three ways:
If it is a leaf node, simply delete it.
If it has a child, have its parent point to its child.
If it has two children, replace it with its next largest value; move to
the right once, and then left as far as you can go. Use the value found
here as its replacement, and then delete that value.

Full disclosure, the logic for this method was supported by a webpage
provided by the Odin Project but converted to Ruby syntax. That being
said, Rubocop complains about line length/complexity for this method, so
perhaps improvements can be made at another time.