Implement #depth for # of edges to target value
Find the node with the given value starting from the root node. This is
more or less like #find but keeps a count along the way. Incidentally,
this is useful for finding the shortest path to a given value, thus it
is likely useful for the future project Knight's Trevails.